Make Entry.Size and ChunkUploadResponse.Offset be int64 - fixes #19
This allows the code to deal with files > 2 GB in size.  Previous to
this change a file > 2 GB would cause a JSON unmarshal error.

This is a breaking change to the API.  However the Go compiler will
throw errors at the right places when compiling so should be easy to
fix up the code.